Fix warnings in JIT test build

Commit migrated from https://github.com/dotnet/coreclr/commit/1115b6f4999d88a48cffd73dbe92cd4252427268

+++ src/coreclr/tests/src/JIT/Directed/pinning/object-pin/mirror.cpp
@@ -4,13 +4,14 @@
 #define EXPORT_API extern "C" __attribute__((visibility("default")))
 #endif 
 
+#include <cstddef>
 
 EXPORT_API unsigned __int32 Ret_Int(unsigned __int32 argVal){
 	unsigned __int32 retVal = (unsigned __int32)argVal;
 	return retVal;
 }
 EXPORT_API unsigned __int32 Ret_Ptr(void *argVal){
-	unsigned __int32 retVal = (unsigned __int32)argVal;
+	unsigned __int32 retVal = (unsigned __int32)(size_t)argVal;
 	return retVal;
 }
